fix datatype

+++ tasks/mpi/sozonov_i_gaussian_method_horizontal_strip_scheme/func_tests/main.cpp
@@ -18,7 +18,7 @@ TEST(sozonov_i_gaussian_method_horizontal_strip_scheme_mpi, test_10_unknowns) {
   std::shared_ptr<ppc::core::TaskData> taskDataPar = std::make_shared<ppc::core::TaskData>();
 
   if (world.rank() == 0) {
-    for (size_t i = 0; i < count; ++i) {
+    for (int i = 0; i < count; ++i) {
       global_mat[i * (count + 1) + i] = 1;
       global_mat[i * (count + 1) + count] = i + 1;
     }
@@ -71,7 +71,7 @@ TEST(sozonov_i_gaussian_method_horizontal_strip_scheme_mpi, test_50_unknowns) {
   std::shared_ptr<ppc::core::TaskData> taskDataPar = std::make_shared<ppc::core::TaskData>();
 
   if (world.rank() == 0) {
-    for (size_t i = 0; i < count; ++i) {
+    for (int i = 0; i < count; ++i) {
       global_mat[i * (count + 1) + i] = 1;
       global_mat[i * (count + 1) + count] = i + 1;
     }
@@ -124,7 +124,7 @@ TEST(sozonov_i_gaussian_method_horizontal_strip_scheme_mpi, test_100_unknowns) {
   std::shared_ptr<ppc::core::TaskData> taskDataPar = std::make_shared<ppc::core::TaskData>();
 
   if (world.rank() == 0) {
-    for (size_t i = 0; i < count; ++i) {
+    for (int i = 0; i < count; ++i) {
       global_mat[i * (count + 1) + i] = 1;
       global_mat[i * (count + 1) + count] = i + 1;
     }
@@ -177,7 +177,7 @@ TEST(sozonov_i_gaussian_method_horizontal_strip_scheme_mpi, test_200_unknowns) {
   std::shared_ptr<ppc::core::TaskData> taskDataPar = std::make_shared<ppc::core::TaskData>();
 
   if (world.rank() == 0) {
-    for (size_t i = 0; i < count; ++i) {
+    for (int i = 0; i < count; ++i) {
       global_mat[i * (count + 1) + i] = 1;
       global_mat[i * (count + 1) + count] = i + 1;
     }
